+++ iam-login-service/src/main/java/it/infn/mw/iam/audit/events/tokens/RegistrationTokenIssuedEvent.java
@@ -1,0 +1,28 @@
+/**
+ * Copyright (c) Istituto Nazionale di Fisica Nucleare (INFN). 2016-2021
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package it.infn.mw.iam.audit.events.tokens;
+
+import org.mitre.oauth2.model.OAuth2AccessTokenEntity;
+
+public class RegistrationTokenIssuedEvent extends SignedTokenEvent {
+
+  private static final long serialVersionUID = 1L;
+
+  public RegistrationTokenIssuedEvent(Object source, OAuth2AccessTokenEntity registrationToken) {
+    super(source, registrationToken.getJwt(), registrationToken.getAuthenticationHolder(), "Issue registration token");
+  }
+
+}
