+++ iam-login-service/src/main/java/it/infn/mw/iam/config/IamProperties.java
@@ -19,6 +19,9 @@ import java.util.EnumMap;
 import java.util.List;
 import java.util.Map;
 
+import javax.validation.constraints.NotBlank;
+import javax.validation.constraints.Pattern;
+
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.stereotype.Component;
 import org.springframework.validation.annotation.Validated;
@@ -709,6 +712,10 @@ public class IamProperties {
 
   private String host;
 
+  @Pattern(
+      regexp = ".*/",
+      message = "issuer must end with '/'"
+  )
   private String issuer;
 
   private String baseUrl;
